feat: [unify-dual-work-failure-json-fields-004] - Populate provider_failure at event emission boundary
