+++ pkg/replay/event_log_test.go
@@ -127,9 +127,7 @@ func replayDispatchCompletedEvent(t *testing.T, completionID string, result inte
 		OutputWork:      generatedReplayOutputWorkPtr(result.RecordedOutputWork),
 		Error:           stringPtrIfNotEmpty(result.Error),
 		Feedback:        stringPtrIfNotEmpty(result.Feedback),
-		ProviderFailure: interfaces.GeneratedWorkFailureMetadata(
-			interfaces.CanonicalWorkFailureMetadata(result.FailureMetadata, result.ProviderFailure),
-		),
+		ProviderFailure: interfaces.PublishedProviderFailureMetadata(result.FailureMetadata, result.ProviderFailure),
 		Metrics:         generatedWorkMetrics(result.Metrics),
 	}
 	var union factoryapi.FactoryEvent_Payload
